requirements.txt
django
djangorestframework 
djangorestframework-simplejwt
django-rest-framework-docs
pandas
numpy
pymysql
sqlalchemy